backend/requirements.txt
django
djangorestframework
django-filter
django-cors-headers
psycopg2-binary
pillow
autoflake
gunicorn
python-dotenv
django-imagekit
